move algolpage to ui
