+++ modules/next/commands/helpers/makeGameRulesPage.ts
@@ -11,7 +11,7 @@ import React, { Fragment } from "react";
 
 import Head from "next/head";
 import { PayloadArticlePage } from "../../../../../ui/src/components/PayloadArticlePage";
-import { AlgolPage } from "../../../../helpers/pageProps";
+import { AlgolPage } from "../../../../../ui/src/helpers/algolPage";
 import article from "../../../../../payloads/dist/articles/gamesRules/${meta.id}";
 
 export const GameRulePage: AlgolPage = props => {
