+++ modules/next/helpers/index.ts
@@ -1,3 +1,2 @@
 export * from "./appActions";
 export * from "./useBattleNavActions";
-export * from "./pageProps";
