+++ modules/next/pages/tags/irreversible/index.tsx
@@ -4,7 +4,7 @@ import React, { Fragment } from "react";
 
 import Head from "next/head";
 import { PayloadArticlePage } from "../../../../ui/src/components/PayloadArticlePage";
-import { AlgolPage } from "../../../helpers/pageProps";
+import { AlgolPage } from "../../../../ui/src/helpers/algolPage";
 
 import article from "../../../../payloads/dist/articles/tags/irreversible";
